@@ -15,7 +15,7 @@ import {
   Divider,
   Input,
   LabeledList,
-  NumberInput,
+  Slider,
   Section,
   Stack,
   Tabs,
@@ -39,6 +39,7 @@ import {
   selectHighlightSettings,
   selectHighlightSettingById,
 } from './selectors';
+import { SettingsStatPanel } from './SettingsStatPanel';
 
 export const SettingsPanel = (props, context) => {
   const activeTab = useSelector(context, selectActiveTab);
@@ -70,6 +71,7 @@ export const SettingsPanel = (props, context) => {
         {activeTab === 'general' && <SettingsGeneral />}
         {activeTab === 'chatPage' && <ChatPageSettings />}
         {activeTab === 'textHighlight' && <TextHighlightSettings />}
+        {activeTab === 'statPanel' && <SettingsStatPanel />}
       </Stack.Item>
     </Stack>
   );
@@ -105,7 +107,7 @@ export const SettingsGeneral = (props, context) => {
           </LabeledList.Item>
           <LabeledList.Item label="Font style">
             <Stack.Item>
-              {(!freeFont && (
+              {!freeFont ? (
                 <Collapsible
                   title={fontFamily}
                   width={'100%'}
@@ -137,7 +139,7 @@ export const SettingsGeneral = (props, context) => {
                     />
                   ))}
                 </Collapsible>
-              )) || (
+              ) : (
                 <Stack>
                   <Input
                     width={'100%'}
@@ -164,13 +166,12 @@ export const SettingsGeneral = (props, context) => {
             </Stack.Item>
           </LabeledList.Item>
           <LabeledList.Item label="Font size">
-            <NumberInput
-              width="4.2em"
+            <Slider
               step={1}
-              stepPixelSize={10}
+              stepPixelSize={17.5}
               minValue={8}
-              maxValue={32}
               value={fontSize}
+              maxValue={32}
               unit="px"
               format={(value) => toFixed(value)}
               onChange={(e, value) =>
@@ -183,13 +184,11 @@ export const SettingsGeneral = (props, context) => {
             />
           </LabeledList.Item>
           <LabeledList.Item label="Line height">
-            <NumberInput
-              width="4.2em"
+            <Slider
               step={0.01}
-              stepPixelSize={2}
               minValue={0.8}
-              maxValue={5}
               value={lineHeight}
+              maxValue={5}
               format={(value) => toFixed(value, 2)}
               onDrag={(e, value) =>
                 dispatch(
